Add mouse mode Off
This adds the mouse event report mode Off which is the new default mode.
Previously, Click was the default mode which caused that after the program exited,
the Click mode was still active and reported mouse events, because the at_exit handler wasn't ran.
Now only one at_exit handler will be registered which disables the current mode and resets the mouse protocol.